+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.4] - 2021-01-27
 ### Added
+- Recursively clone git submodules.
 - Improve proxy setup.
 - Detect proxy config.
 - Upgrade procedure from old directory schema to the new one with all the fwcloud processes under /opt/fwcloud/